ai-service/app.py
from flask import Flask
from dotenv import load_dotenv

from routes.health import health_bp
from routes.describe import describe_bp
from routes.recommend import recommend_bp
from routes.report import report_bp
from routes.search import search_bp

load_dotenv()

app = Flask(__name__)

app.register_blueprint(health_bp)
app.register_blueprint(describe_bp)
app.register_blueprint(recommend_bp)
app.register_blueprint(report_bp)
app.register_blueprint(search_bp)

if __name__ == "__main__":
    app.run(debug=True, port=5000)