ai-service/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==2.7.0
chromadb==0.5.5
redis==5.0.4